+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+
+## 2023-10-31
+### Added
+- 'TinkoffHistoryData' data provider for data downloaded from 'https://invest-public-api.tinkoff.ru/history-data'
+[More information](https://tinkoff.github.io/investAPI/get_history/)
+
+
 ## 2022-11-02
 ### Changed
 - The project has been moved to separate repository